Move createCredentialV1 + improve the related test
