+++ packages/common/src/identity.ts
@@ -1,14 +1,11 @@
 import * as wasm from '@concordium/rust-bindings';
 import {
     ArInfo,
-    AttributeKey,
-    CredentialDeploymentInfo,
     CryptographicParameters,
+    IdObjectRequestV1,
     IpInfo,
-    SignedCredentialDeploymentDetails,
     Versioned,
 } from './types';
-import { TransactionExpiry } from './types/transactionExpiry';
 
 export type IdentityRequestInput = {
     ipInfo: IpInfo;
@@ -19,24 +16,6 @@ export type IdentityRequestInput = {
     identityIndex: number;
     arThreshold: number;
 };
-
-export type IpArData = {
-    encPrfKeyShare: string;
-    proofComEncEq: string;
-};
-
-export interface IdObjectRequestV1 {
-    idCredPub: string;
-    choiceArData: {
-        arIdentities: number[];
-        threshold: number;
-    };
-    ipArData: Record<string, IpArData>;
-    idCredSecCommitment: string;
-    prfKeyCommitmentWithIP: string;
-    prfKeySharingCoeffCommitments: string[];
-    proofsOfKnowledge: string;
-}
 
 /**
  * Creates a V1 identityRequest.
@@ -50,49 +29,4 @@ export function createIdentityRequest(
     } catch (e) {
         throw new Error(rawRequest);
     }
-}
-
-export interface AttributeList {
-    validTo: string; // "YYYYMM"
-    createdAt: string; // "YYYYMM"
-    maxAccounts: number;
-    chosenAttributes: Record<AttributeKey, string>;
-}
-
-export type IdentityObject = {
-    preIdentityObject: IdObjectRequest;
-    attributeList: AttributeList;
-    signature: string;
-};
-
-export type CredentialInput = {
-    ipInfo: IpInfo;
-    globalContext: CryptographicParameters;
-    arsInfos: Record<string, ArInfo>;
-    idObject: IdentityObject;
-    revealedAttributes: AttributeKey[];
-    seed: string;
-    net: 'Testnet' | 'Mainnet';
-    identityIndex: number;
-    credNumber: number;
-    expiry: number;
-};
-
-/**
- * Creates a V1 credential for a new account.
- */
-export function createCredentialV1(
-    input: CredentialInput
-): SignedCredentialDeploymentDetails {
-    const rawRequest = wasm.createCredentialV1(JSON.stringify(input));
-    let info: CredentialDeploymentInfo;
-    try {
-        info = JSON.parse(rawRequest);
-    } catch (e) {
-        throw new Error(rawRequest);
-    }
-    return {
-        expiry: TransactionExpiry.fromEpochSeconds(BigInt(input.expiry)),
-        cdi: info,
-    };
 }
