+++ packages/common/src/serializationHelpers.ts
@@ -2,7 +2,6 @@ import { Buffer } from 'buffer/';
 import { VerifyKey } from '.';
 import {
     ConfigureDelegationPayload,
-    CredentialDeploymentInfo,
     DelegationTarget,
     DelegationTargetType,
     ParameterType,
